+++ studio-server/src/main/java/com/coremedia/labs/plugins/adapters/cloudinary/CloudinaryContentHubTransformer.java
@@ -3,7 +3,6 @@ package com.coremedia.labs.plugins.adapters.cloudinary;
 import com.coremedia.cap.common.Blob;
 import com.coremedia.contenthub.api.*;
 import com.coremedia.cotopaxi.common.blobs.BlobServiceImpl;
-import com.coremedia.labs.plugins.adapters.cloudinary.rest.CloudinaryAsset;
 import com.coremedia.mimetype.TikaMimeTypeService;
 import com.coremedia.util.TempFileFactory;
 import edu.umd.cs.findbugs.annotations.NonNull;
@@ -11,19 +10,22 @@ import edu.umd.cs.findbugs.annotations.Nullable;
 import org.apache.commons.io.FilenameUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.web.util.HtmlUtils;
 
 import javax.activation.MimeTypeParseException;
 import java.io.IOException;
 import java.io.InputStream;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Objects;
+import java.util.*;
 
 class CloudinaryContentHubTransformer implements ContentHubTransformer {
   private static final Logger LOG = LoggerFactory.getLogger(CloudinaryContentHubTransformer.class);
 
+  private final CloudinaryImportOptions importOptions;
+
   private TikaMimeTypeService tikaMimeTypeService;
+
+  public CloudinaryContentHubTransformer(@NonNull CloudinaryImportOptions importOptions) {
+    this.importOptions = importOptions;
+  }
 
   @Override
   @NonNull
@@ -56,7 +58,11 @@ class CloudinaryContentHubTransformer implements ContentHubTransformer {
   private ContentModel transformCloudinaryItem(CloudinaryItem item) {
     String contentName = FilenameUtils.removeExtension(item.getName());
     ContentModel contentModel = ContentModel.createContentModel(contentName, item.getId(), item.getCoreMediaContentType());
+
     contentModel.put("title", contentName);
+    String importPublicIdAs = importOptions.getImportPublicIdAs();
+    if (importPublicIdAs != null && !importPublicIdAs.isEmpty())
+      contentModel.put(importPublicIdAs, item.getAsset().getPublicId());
 
     String type = item.getCoreMediaContentType();
     Map<String, Object> additionalProps = new HashMap<>();
@@ -65,10 +71,8 @@ class CloudinaryContentHubTransformer implements ContentHubTransformer {
       additionalProps.putAll(Objects.requireNonNull(getMediaProperties(item)));
     }
 
-    String description = extractDescription(item.getAsset());
-    if (description != null) {
-      contentModel.put("detailText", ContentCreationUtil.convertStringToRichtext(description));
-    }
+    // example how to populate a Richtext property
+    // contentModel.put("detailText", ContentCreationUtil.convertStringToRichtext(description));
 
     //add additional properties
     additionalProps.forEach(contentModel::put);
@@ -77,22 +81,24 @@ class CloudinaryContentHubTransformer implements ContentHubTransformer {
   }
 
   @Nullable
-  private String extractDescription(@Nullable CloudinaryAsset asset) {
-    return asset == null ? null : HtmlUtils.htmlUnescape(asset.getFolder());
-  }
-
-  @Nullable
   private Map<String, Object> getMediaProperties(CloudinaryItem item) {
-
     Map<String, Object> result = new HashMap<>();
-
     try {
-      InputStream stream = item.stream();
-      BlobServiceImpl blobService = new BlobServiceImpl(new TempFileFactory(), getTika());
-      Blob blob = blobService.fromInputStream(stream, getTika().getMimeTypeForResourceName(item.getName() + "." + item.getFormat()));
-      result.put("data", blob);
-      if (stream != null) {
-        stream.close();
+      // handling for Videos import options
+      String type = item.getCoreMediaContentType();
+      // check if we are importing the blob
+      if (!type.equals("CMVideo") || importOptions.isImportVideoBlob()) {
+        InputStream stream = item.stream(true);
+        BlobServiceImpl blobService = new BlobServiceImpl(new TempFileFactory(), getTika());
+        Blob blob = blobService.fromInputStream(stream, getTika().getMimeTypeForResourceName(item.getName() + "." + item.getFormat()));
+        result.put("data", blob);
+        if (stream != null) {
+          stream.close();
+        }
+      }
+      // check if we are importing the URL
+      if(type.equals("CMVideo") && importOptions.isImportVideoURL()) {
+        result.put("dataUrl", item.getAsset().getSecureUrl());
       }
       return result;
     } catch (IOException | MimeTypeParseException e) {